Safari-609.1.11.1/LayoutTests/js/dom/exceptions-thrown-in-callbacks-expected.txt

CONSOLE MESSAGE: line 8: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 17: PASS: toString called on exception value thrown from event handler
CONSOLE MESSAGE: line 23: PASS: toString called on exception value thrown from sql transaction callback
CONSOLE MESSAGE: line 33: PASS: toString called on exception value thrown from sql error callback
CONSOLE MESSAGE: line 41: PASS: toString called on exception value thrown from timer
This test ensures that exceptions are handled correctly by the various callback mechanisms present in WebCore.
